+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Provide code navigation support for `macro` and `function` in `.ftl` files.
 - Provide existed `macro` completion items.
 - Provide completions items for all `FreeMarker` directives.
+- Provide hover tips for all `FreeMarker` directives.